backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
biopython==1.79
rdkit==2024.9.5